2005-10-29  Stefan Jahn  <stefan@lkcc.org>
        * check_netlist.cpp: Skip range test if a property is assigned
        by a (sweep) variable.

        * precision.h: Improved --enable-double behaviour by
        introducing double representation dependent precision
        constants.

2005-10-27  Stefan Jahn  <stefan@lkcc.org>

        * parse_netlist.y: Fixed smaller memory leaks in the netlist
        parser.

        * vector.cpp, equation.cpp, evaluate.cpp, net.cpp: Fixed loads
        of memory leaks.

        * environment.cpp: Fixed memory leak (deleting substrates
        and constants).

        * dataset.cpp (load_touchstone), check_touchstone.cpp: Fixed
        scanner memory leak.

        * check_netlist.cpp (netlist_destroy): Calling
        netlist_lex_destroy() fixes a memory leak in the netlist
        scanner.

+++ qucs-core/src/sweep.cpp
@@ -18,7 +18,7 @@
  * the Free Software Foundation, Inc., 51 Franklin Street - Fifth Floor,
  * Boston, MA 02110-1301, USA.  
  *
- * $Id: sweep.cpp,v 1.7 2005/06/02 18:17:51 raimi Exp $
+ * $Id: sweep.cpp,v 1.8 2005/10/31 16:15:31 ela Exp $
  *
  */
 
@@ -113,7 +113,7 @@ char * sweep::toString (void) {
   strcpy (txt, "[");
   for (int i = 0; i < size; i++) {
     static char str[256];  // enough for a real number
-    sprintf (str, "%g", get (i));
+    sprintf (str, "%g", (double) get (i));
     txt = (char *) realloc (txt, len += strlen (str));
     strcat (txt, str);
     if (i != size - 1) strcat (txt, ";");
